Implement server info handler as route endpoint
