+++ tests/test_lizmap_api.py
@@ -38,31 +38,31 @@ def test_lizmap_api_default_path_resolution(rootdir: Path):
 def test_lizmap_api_routes(client):
     from lizmap_server.api import errors, routes
 
-    route, match_infos  = routes.find_route(QgsServerRequest.GetMethod, "/projects/list/")
+    route, match_infos  = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/list/")
     assert route is not None
     assert match_infos.get("PATH") == ""
 
-    route, match_infos  = routes.find_route(QgsServerRequest.GetMethod, "/projects/list/foo/bar")
+    route, match_infos  = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/list/foo/bar")
     assert route is not None
     assert match_infos.get("PATH") == "foo/bar"
 
-    route, _ = routes.find_route(QgsServerRequest.GetMethod, "/projects/description")
+    route, _ = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/description")
     assert route is not None
 
     with pytest.raises(errors.HTTPNotFound):
-        route, _ = routes.find_route(QgsServerRequest.GetMethod, "/projects/description/foo")
+        route, _ = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/description/foo")
 
     with pytest.raises(errors.HTTPNotFound):
-        route, _ = routes.find_route(QgsServerRequest.GetMethod, "/projects/layers/")
+        route, _ = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/layers/")
 
-    route, match_infos = routes.find_route(QgsServerRequest.GetMethod, "/projects/layers/bar")
+    route, match_infos = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/layers/bar")
     assert route is not None
     assert match_infos.get("Id") == "bar"
 
-    route, _ = routes.find_route(QgsServerRequest.GetMethod, "/projects/layouts/")
+    route, _ = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/layouts/")
     assert route is not None
 
-    route, match_infos = routes.find_route(QgsServerRequest.GetMethod, "/projects/layouts/foobar")
+    route, match_infos = routes.find_route(QgsServerRequest.GetMethod, "/api/v1/projects/layouts/foobar")
     assert route is not None
     assert match_infos.get("Name") == "foobar"
 
@@ -71,7 +71,7 @@ def test_lizmap_api_projects_list(client):
     """Test the Lizmap API for server settings"""
 
     # The query must work
-    rv = client.get("/api/v1/projects/list/data")
+    rv = client.get("/lizmap/api/v1/projects/list/data")
     assert rv.status_code == 200
     assert rv.headers.get("Content-Type", "").find("application/json") == 0
 
@@ -82,7 +82,7 @@ def test_lizmap_api_projects_list(client):
 def test_lizmap_api_projects_description(client):
     """Test the Lizmap API for server settings"""
 
-    rv = client.get("/api/v1/projects/description?p=/data/montpellier/montpellier.qgs")
+    rv = client.get("/lizmap/api/v1/projects/description?p=/data/montpellier/montpellier.qgs")
     assert rv.status_code == 200
     assert rv.headers.get("Content-Type", "").find("application/json") == 0
 
@@ -95,7 +95,7 @@ def test_lizmap_api_projects_layers(client):
 
     layer_id = "tram_stop_work20150416102656130"
 
-    rv = client.get(f"/api/v1/projects/layers/{layer_id}?p=/data/montpellier/montpellier.qgs")
+    rv = client.get(f"/lizmap/api/v1/projects/layers/{layer_id}?p=/data/montpellier/montpellier.qgs")
     assert rv.status_code == 200
     assert rv.headers.get("Content-Type", "").find("application/json") == 0
 
@@ -107,14 +107,14 @@ def test_lizmap_api_projects_layers(client):
 def test_lizmap_api_projects_layouts(client):
     """Test the Lizmap API for server settings"""
 
-    rv = client.get("/api/v1/projects/layouts/?p=/data/montpellier/montpellier.qgs")
+    rv = client.get("/lizmap/api/v1/projects/layouts/?p=/data/montpellier/montpellier.qgs")
     assert rv.status_code == 200
     assert rv.headers.get("Content-Type", "").find("application/json") == 0
 
     content = json.loads(rv.content.decode("utf-8"))
     print("\n::test_lizmap_server_api::layouts::\n", content)
 
-    rv = client.get("/api/v1/projects/layouts/Landscape A4?p=/data/montpellier/montpellier.qgs")
+    rv = client.get("/lizmap/api/v1/projects/layouts/Landscape A4?p=/data/montpellier/montpellier.qgs")
     assert rv.status_code == 200
     assert rv.headers.get("Content-Type", "").find("application/json") == 0
